Fix gas unit for SMETS1 meters
SMETS1 meters report gas consumption in kWh, see
https://developer.octopus.energy/docs/api/#consumption

This commit adds an option to set the gas meter type should a SMETS1
meter be in use and the readings be in kWh.

The default behaviour still assumes SMETS2 and assumes the readings to
be in m3.

(I believe it should be possible to determine the meter type from the
serial number, avoiding the need for the extra config option, but I
could not find concrete information about this for gas meters.)